refactor code consistency #112
