+++ src/eventHandlers/index.ts
@@ -3,4 +3,4 @@ export * from "./auctions"
 export * from "./balances"
 export * from "./marketplaces"
 export * from "./nfts"
-export * from "./nftTransfer"
+export * from "./nftOperations"
